chore_: code review - move to separate file and use AES encryption
