+++ protocol/messenger_backup_test.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"os"
-	"path/filepath"
 	"reflect"
 	"testing"
 	"time"
@@ -15,7 +13,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/event"
 
-	"github.com/status-im/status-go/params"
 	v1protocol "github.com/status-im/status-go/protocol/v1"
 	"github.com/status-im/status-go/protocol/wakusync"
 	"github.com/status-im/status-go/services/accounts/accountsevent"
@@ -111,76 +108,6 @@ func (s *MessengerBackupSuite) TestBackupContacts() {
 	s.Require().NoError(err)
 	s.Require().NotEmpty(lastBackup)
 	s.Require().Equal(clock, lastBackup)
-}
-
-func (s *MessengerBackupSuite) TestBackupContactsLocally() {
-	backupOptions := []Option{
-		WithLocalBackup(&params.BackupConfig{
-			DataDir: filepath.Join(s.tmpdir, params.BackupsRelativePath),
-		}),
-	}
-
-	// Create bob1
-	privateKey, err := crypto.GenerateKey()
-	s.Require().NoError(err)
-	bob1, err := newMessengerWithKey(s.shh, privateKey, s.logger, backupOptions)
-	s.Require().NoError(err)
-	defer TearDownMessenger(&s.Suite, bob1)
-
-	// Create bob2
-	bob2, err := newMessengerWithKey(s.shh, bob1.identity, s.logger, backupOptions)
-	s.Require().NoError(err)
-	defer TearDownMessenger(&s.Suite, bob2)
-
-	// Make sure there is no backup at first
-	backupFile := filepath.Join(bob1.config.backupConfig.DataDir, "user_data.bkp")
-	err = os.RemoveAll(backupFile)
-	s.Require().NoError(err)
-
-	// Create 2 contacts
-	contact1Key, err := crypto.GenerateKey()
-	s.Require().NoError(err)
-	contactID1 := types.EncodeHex(crypto.FromECDSAPub(&contact1Key.PublicKey))
-
-	_, err = bob1.AddContact(context.Background(), &requests.AddContact{ID: contactID1})
-	s.Require().NoError(err)
-
-	contact2Key, err := crypto.GenerateKey()
-	s.Require().NoError(err)
-	contactID2 := types.EncodeHex(crypto.FromECDSAPub(&contact2Key.PublicKey))
-
-	_, err = bob1.AddContact(context.Background(), &requests.AddContact{ID: contactID2})
-	s.Require().NoError(err)
-
-	s.Require().Len(bob1.Contacts(), 2)
-
-	actualContacts := bob1.Contacts()
-	if actualContacts[0].ID == contactID1 {
-		s.Require().Equal(actualContacts[0].ID, contactID1)
-		s.Require().Equal(actualContacts[1].ID, contactID2)
-	} else {
-		s.Require().Equal(actualContacts[0].ID, contactID2)
-		s.Require().Equal(actualContacts[1].ID, contactID1)
-	}
-
-	s.Require().Equal(ContactRequestStateSent, actualContacts[0].ContactRequestLocalState)
-	s.Require().Equal(ContactRequestStateSent, actualContacts[1].ContactRequestLocalState)
-	s.Require().True(actualContacts[0].added())
-	s.Require().True(actualContacts[1].added())
-
-	// Backup
-	_, err = bob1.BackupData(context.Background())
-	s.Require().NoError(err)
-
-	// Safety check
-	s.Require().Len(bob2.Contacts(), 0)
-
-	// Import the backup file and process it
-	response, err := bob2.importLocalBackupFile(backupFile)
-	s.Require().NoError(err)
-	s.Require().NotNil(response)
-	s.Require().Len(response.Contacts, 2)
-	s.Require().Len(bob2.Contacts(), 2)
 }
 
 func (s *MessengerBackupSuite) TestBackupProfile() {
